Fix initialisation of PWMController and format of twist message
Since the ServoKit initialisation takes a while, any code that comes
after it in the init method isn't run immediately. Since the ros
subscription to the twist topic came before this ServoKit step, the
twist callback was being called even before the PWMController had
finished being initialised, causing errors for the first 10 seconds that
the node was launched.